+++ src/main/java/gdg/farm_in_palm/domain/monitor/dto/MonitorInfoResDTO.java
@@ -1,9 +1,7 @@
 package gdg.farm_in_palm.domain.monitor.dto;
 
 import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 @Getter
 public class MonitorInfoResDTO {
@@ -13,10 +11,10 @@ public class MonitorInfoResDTO {
     private Float humidity;
     private Float groundTemperature;
     private Float groundHumidity;
-    private Integer co2Concentration;
+    private Float co2Concentration;
 
     @Builder
-    public MonitorInfoResDTO(Long monitorId, String monitorName, Float temperature, Float humidity, Float groundTemperature, Float groundHumidity, Integer co2Concentration) {
+    public MonitorInfoResDTO(Long monitorId, String monitorName, Float temperature, Float humidity, Float groundTemperature, Float groundHumidity, Float co2Concentration) {
         this.monitorId = monitorId;
         this.monitorName = monitorName;
         this.temperature = temperature;
